+++ trilby-cli/src/Trilby/Install.hs
@@ -1,6 +1,5 @@
 module Trilby.Install (install) where
 
-import Data.Text qualified as Text
 import Trilby.HNix (writeNixFile)
 import Trilby.Host (Host (Localhost), hostSystem, reboot)
 import Trilby.Install.Config.Host
@@ -11,7 +10,6 @@ import Trilby.Install.Flake
 import Trilby.Install.Options
 import Trilby.System
 import Trilby.Widgets
-import Turtle qualified
 import Prelude
 
 rootMount :: Kernel -> Path Abs Dir
@@ -21,21 +19,21 @@ rootMount Darwin = $(mkAbsDir "/")
 trilbyDir :: Kernel -> Path Abs Dir
 trilbyDir = trilbyHome . rootMount
 
-install :: InstallOpts Maybe -> App ()
+install :: (HasCallStack) => InstallOpts Maybe -> App ()
 install (askOpts -> opts) = do
     system <- hostSystem Localhost
     case system.kernel of
         Linux -> installLinux opts
         Darwin -> installDarwin opts
 
-installLinux :: InstallOpts App -> App ()
+installLinux :: (HasCallStack) => InstallOpts App -> App ()
 installLinux opts | Just FlakeOpts{..} <- opts.flake = do
     let diskoRef = Disko.Flake flakeRef
     formatDisk opts.format diskoRef
     mountRoot diskoRef
     nixosInstall flakeRef
     whenM copyFlake do
-        storePath <- Text.strip <$> shell ("nix flake archive --json " <> flakeRef.url <> " | jq --raw-output .path") empty
+        Just storePath <- shellOutTextFirstLine ["nix flake archive --json " <> flakeRef.url <> " | jq --raw-output .path"]
         asRoot cmd_ ["cp", "-r", storePath, fromPath $ trilbyDir Linux]
         asRoot cmd_ ["chown", "-R", "1000:1000", fromPath $ trilbyDir Linux]
     reboot opts.reboot Localhost
@@ -50,32 +48,28 @@ installLinux opts = withTempFile $(mkRelFile "disko.nix") \diskoFile -> do
     nixosInstall flakeRef
     reboot opts.reboot Localhost
 
-formatDisk :: App Bool -> FileOrFlake -> App ()
+formatDisk :: (HasCallStack) => App Bool -> FileOrFlake -> App ()
 formatDisk f d = whenM f do
-    logWarn "Formatting disk ... "
+    logAttention_ "Formatting disk ... "
     disko $ Format d
 
-mountRoot :: FileOrFlake -> App ()
+mountRoot :: (HasCallStack) => FileOrFlake -> App ()
 mountRoot d = unlessM rootIsMounted do
-    logWarn "Partitions are not mounted"
+    logAttention_ "Partitions are not mounted"
     unlessM (yesNoButtons "Attempt to mount the partitions?" True) $
         errorExit "Cannot install without mounted partitions"
     disko $ Mount d
   where
-    rootIsMounted = (ExitSuccess ==) . fst <$> cmd' ["mountpoint", "-q", fromPath $ rootMount Linux]
+    rootIsMounted = (ExitSuccess ==) <$> cmdCode ["mountpoint", "-q", fromPath $ rootMount Linux]
 
 flakeNix, defaultNix, configurationNix :: Path Rel File
 flakeNix = $(mkRelFile "flake.nix")
 defaultNix = $(mkRelFile "default.nix")
 configurationNix = $(mkRelFile "configuration.nix")
 
-data Owner = Owner {uid :: Int, gid :: Int}
-
-instance Show Owner where
-    show Owner{..} = show uid <> ":" <> show gid
-
 setupHost
-    :: Kernel
+    :: (HasCallStack)
+    => Kernel
     -> InstallOpts App
     -> (Path Rel Dir -> Path Rel Dir -> App ())
     -> App FlakeRef
@@ -85,13 +79,7 @@ setupHost kernel opts actions = do
     release <- opts.release
     realTrilbyDir <- canonicalizePath $ trilbyDir kernel
     inDir realTrilbyDir do
-        owner <-
-            case kernel of
-                Linux -> pure Owner{uid = 1000, gid = 1000}
-                Darwin -> do
-                    uid <- read . fromText <$> cmd ["id", "-u"]
-                    gid <- read . fromText <$> cmd ["id", "-g"]
-                    pure Owner{..}
+        owner <- currentOwner
         asRoot cmd_ ["chown", "-R", ishow owner, fromPath realTrilbyDir]
         writeNixFile flakeNix $ flake kernel release
         hostDir <- parseRelDir . fromText $ "hosts/" <> hostname
@@ -134,7 +122,7 @@ setupHost kernel opts actions = do
                 Linux -> do
                     writeFile $(mkRelFile "hardware-configuration.nix")
                         =<< asRoot
-                            cmd
+                            cmdOutText
                             [ "nixos-generate-config"
                             , "--show-hardware-config"
                             , "--no-filesystems"
@@ -150,28 +138,28 @@ setupHost kernel opts actions = do
         actions hostDir userDir
         whenM opts.edit do
             editor <- fromMaybe "nano" <$> lookupEnv "EDITOR"
-            rawCmd_
+            cmd_
                 [ fromString editor
                 , "flake.nix"
                 , fromPath $ hostDir </> configurationNix
                 ]
     pure FlakeRef{url = fromPath realTrilbyDir, output = pure hostname}
 
-hashedPassword :: Text -> App Password
-hashedPassword plain = HashedPassword . firstLine <$> Turtle.strict (Turtle.inproc "mkpasswd" [plain] empty)
+hashedPassword :: (HasCallStack) => Text -> App Password
+hashedPassword plain = maybe (fail "mkpasswd failed") (pure . HashedPassword) =<< cmdOutTextFirstLine ["mkpasswd", plain]
 
-nixosInstall :: FlakeRef -> App ()
+nixosInstall :: (HasCallStack) => FlakeRef -> App ()
 nixosInstall flakeRef = do
-    logWarn "Performing installation ... "
+    logAttention_ "Performing installation ... "
     -- TODO(vkleen): this shouldn't work and neither should it be necessary ...
-    (withTrace . asRoot) rawCmd_ . sconcat $
+    (withTrace . asRoot) cmd_ . sconcat $
         [ ["nix", "build"]
         , ["--store", "/mnt"]
         , ["--impure"]
         , ["--accept-flake-config"]
         , ["trilby#nix-monitored"]
         ]
-    (withTrace . asRoot) rawCmd_ . sconcat $
+    (withTrace . asRoot) cmd_ . sconcat $
         [ ["nixos-install"]
         , ["--flake", ishow flakeRef]
         , ["--option", "accept-flake-config", "true"]
@@ -179,7 +167,7 @@ nixosInstall flakeRef = do
         , ["--impure"]
         ]
 
-installDarwin :: InstallOpts App -> App ()
+installDarwin :: (HasCallStack) => InstallOpts App -> App ()
 -- installDarwin opts | Just FlakeOpts{..} <- opts.flake = pure ()
 installDarwin opts = do
     flakeRef <- setupHost Darwin opts \_ _ -> pure ()