+++ trilby-cli/src/Trilby/Widgets.hs
@@ -11,7 +11,6 @@ import Prelude
 
 textInputOpts :: Bool -> Bool -> Text -> Text -> App Text
 textInputOpts multiline required ((<> " ") -> prompt) (RopeZipper.fromText -> value) = do
-    logDebug prompt
     text <- runWidgetIO TextInput{valueTransform = id, ..}
     pure $ RopeZipper.toText text.value
 
@@ -21,7 +20,7 @@ textInput = textInputOpts False True
 multilineTextInput :: Text -> Text -> App Text
 multilineTextInput = textInputOpts True False
 
-passwordInput :: Text -> App Text
+passwordInput :: (HasCallStack) => Text -> App Text
 passwordInput ((<> " ") -> prompt) = do
     let getPw :: TextInput -> App Text
         getPw = fmap (RopeZipper.toText . (.value)) . runWidgetIO
@@ -38,12 +37,11 @@ passwordInput ((<> " ") -> prompt) = do
     if pw1 == pw2
         then pure pw1
         else do
-            logError "Passwords do not match"
+            logAttention_ "Passwords do not match"
             passwordInput prompt
 
-buttons :: (Eq a, Show a) => Text -> [(a, Char)] -> Int -> (a -> Text) -> App a
+buttons :: (HasCallStack, Eq a, Show a) => Text -> [(a, Char)] -> Int -> (a -> Text) -> App a
 buttons prompt values selected buttonText = do
-    logDebug prompt
     b <-
         runWidgetIO
             Buttons
@@ -54,16 +52,16 @@ buttons prompt values selected buttonText = do
     case values !? b.selected of
         Just (a, _) -> pure a
         Nothing -> do
-            logError "Invalid selection"
+            logAttention_ "Invalid selection"
             buttons prompt values selected buttonText
 
-yesNoButtons :: Text -> Bool -> App Bool
+yesNoButtons :: (HasCallStack) => Text -> Bool -> App Bool
 yesNoButtons prompt defaultValue = do
     let values = [("Yes", 'Y'), ("No", 'N')] :: [(Text, Char)]
     let selected = if defaultValue then 0 else 1
     ("Yes" ==) <$> buttons prompt values selected id
 
-multiSelect :: (Eq a, Show a) => Text -> [a] -> [a] -> (a -> Text) -> Int -> Int -> App [a]
+multiSelect :: (HasCallStack, Eq a, Show a) => Text -> [a] -> [a] -> (a -> Text) -> Int -> Int -> App [a]
 multiSelect prompt options selections optionText minSelect maxSelect
     | minSelect > maxSelect = error "multiSelect: minSelect must be < maxSelect"
     | length options < minSelect = error "multiSelect: called with fewer options than minSelect"
@@ -82,28 +80,28 @@ multiSelect prompt options selections optionText minSelect maxSelect
                 }
             <&> (fmap (.value) . filter (.checked) . (.options))
 
-multiSelect1 :: (Eq a, Show a) => Text -> [a] -> [a] -> (a -> Text) -> Int -> Int -> App (NonEmpty a)
+multiSelect1 :: (HasCallStack, Eq a, Show a) => Text -> [a] -> [a] -> (a -> Text) -> Int -> Int -> App (NonEmpty a)
 multiSelect1 prompt options selections optionText minSelect maxSelect
     | minSelect < 1 = errorExit "multiSelect1: minSelect must be > 0"
     | otherwise = NonEmpty.fromList <$> multiSelect prompt options selections optionText minSelect maxSelect
 
-multiSelectEnum :: (Eq a, Bounded a, Enum a, Show a) => Text -> [a] -> Int -> Int -> App [a]
+multiSelectEnum :: (HasCallStack, Eq a, Bounded a, Enum a, Show a) => Text -> [a] -> Int -> Int -> App [a]
 multiSelectEnum prompt selections = multiSelect prompt [minBound .. maxBound] selections ishow
 
-multiSelectEnum1 :: (Eq a, Bounded a, Enum a, Show a) => Text -> [a] -> Int -> Int -> App (NonEmpty a)
+multiSelectEnum1 :: (HasCallStack, Eq a, Bounded a, Enum a, Show a) => Text -> [a] -> Int -> Int -> App (NonEmpty a)
 multiSelectEnum1 prompt selections = multiSelect1 prompt [minBound .. maxBound] selections ishow
 
-select :: (Eq a, Show a) => Text -> [a] -> Maybe a -> (a -> Text) -> App a
+select :: (HasCallStack, Eq a, Show a) => Text -> [a] -> Maybe a -> (a -> Text) -> App a
 select prompt options selection optionText
     | [o] <- options = pure o
     | null options, Just o <- selection = pure o
     | null options = errorExit "select: called with no options and no default value"
     | otherwise = NonEmpty.head <$> multiSelect1 prompt options (maybeToList selection) optionText 1 1
 
-selectEnum :: (Eq a, Bounded a, Enum a, Show a) => Text -> Maybe a -> App a
+selectEnum :: (HasCallStack, Eq a, Bounded a, Enum a, Show a) => Text -> Maybe a -> App a
 selectEnum prompt defaultValues = select prompt [minBound .. maxBound] defaultValues ishow
 
-searchSelect :: (Eq a, Show a) => Text -> [a] -> [a] -> [a] -> (a -> Text) -> Int -> Int -> App [a]
+searchSelect :: (HasCallStack, Eq a, Show a) => Text -> [a] -> [a] -> [a] -> (a -> Text) -> Int -> Int -> App [a]
 searchSelect prompt options visibleOptions selections optionText minSelect maxSelect
     | minSelect > maxSelect = error "searchSelect: minSelect must be < maxSelect"
     | length options < minSelect = errorExit "searchSelect: need at least as many options as minSelect"
@@ -126,7 +124,11 @@ searchSelect prompt options visibleOptions selections optionText minSelect maxSe
                 }
             <&> (.selections)
 
-searchSelect1 :: (Eq a, Show a) => Text -> [a] -> [a] -> [a] -> (a -> Text) -> Int -> Int -> App (NonEmpty a)
-searchSelect1 ((<> " ") -> prompt) options visibleOptions selections optionText minSelect maxSelect
+searchSelectSome :: (HasCallStack, Eq a, Show a) => Text -> [a] -> [a] -> [a] -> (a -> Text) -> Int -> Int -> App (NonEmpty a)
+searchSelectSome ((<> " ") -> prompt) options visibleOptions selections optionText minSelect maxSelect
     | minSelect < 1 = errorExit "searchSelect1: minSelect must be > 0"
     | otherwise = NonEmpty.fromList <$> searchSelect prompt options visibleOptions selections optionText minSelect maxSelect
+
+searchSelect1 :: (HasCallStack, Eq a, Show a) => Text -> [a] -> [a] -> [a] -> (a -> Text) -> App a
+searchSelect1 ((<> " ") -> prompt) options visibleOptions selections optionText =
+    NonEmpty.head <$> searchSelectSome prompt options visibleOptions selections optionText 1 1